Move code order and un-comment code
Moved code order in an attempt to get the game to log a win when the
winning cell is clicked, instead of when a subsequent cell is clicked
(but this may be an issue to solve with ajax, and or through the api).

un-commented code to get the files talking to each other.